Accept  Store Draw Registration And Group Registration From User Without Authentication.

+++ lib/models/competitions/competition.dart
@@ -13,7 +13,7 @@ class Competition {
 
   bool isLive;
   bool isOver;
-  double joiningFee;
+  int joiningFee;
 
   SectionName competitionSectionName;
